requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Frontend Dependencies  
streamlit==1.29.0
streamlit-chat==0.1.1

# Document Processing
langchain==0.0.350
langchain-community==0.0.10
pypdf==3.17.1
python-docx==1.1.0
unstructured==0.11.6
markdown==3.5.1

# Vector Database & Embeddings
chromadb==0.4.18
sentence-transformers==2.7.0
huggingface-hub>=0.19.0
faiss-cpu==1.7.4

# LLM Integration
litellm==1.17.0
openai==1.6.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.0
numpy==1.24.3
pandas==2.1.4
pyyaml==6.0.1